FIXME: add clue sets + SPLIT a part and fixup to PLI shards. It makes a pli class a struct instead of a class
Not building, just static fields filling: Packs of predicates
and correction map

Need to fix the TODO's, and properly implement PredicateSet, as this
is used there (commented so far)

Also need to change test later as we will add build_clues virtual
function, I guess. Need to think about it.